add poetry.ts
