+++ frontend/src/scripts/test/test-logic.js
@@ -24,10 +24,10 @@ import * as TestTimer from "./test-timer";
 import * as OutOfFocus from "./out-of-focus";
 import * as AccountButton from "../elements/account-button";
 import * as DB from "../db";
-import * as Replay from "./replay.js";
+import * as Replay from "./replay";
 import axiosInstance from "../axios-instance";
-import * as Poetry from "./poetry.js";
-import * as Wikipedia from "./wikipedia.js";
+import * as Poetry from "./poetry";
+import * as Wikipedia from "./wikipedia";
 import * as TodayTracker from "./today-tracker";
 import * as WeakSpot from "./weak-spot";
 import * as Wordset from "./wordset";
